Added wip.php: moved the messy php code from index.php into this file.
